🧬 Migrate the data model *APIConfig* to module *api_config.api_config*.
