+++ pymock_api/command/check/component.py
@@ -7,12 +7,12 @@ from typing import Any, Callable, Optional
 
 from ..._utils.api_client import URLLibHTTPClient
 from ...model import (
+    APIConfig,
     SubcmdCheckArguments,
     SwaggerConfig,
     deserialize_swagger_api_config,
     load_config,
 )
-from ...model.api_config import APIConfig
 from ...model.api_config import APIParameter as MockedAPIParameter
 from ...model.api_config import ResponseProperty
 from ...model.enums import ResponseStrategy
