+++ pymock_api/model/swagger_config.py
@@ -5,7 +5,7 @@ from abc import ABCMeta, abstractmethod
 from pydoc import locate
 from typing import Any, Dict, List, Optional, Union
 
-from pymock_api.model.api_config import APIConfig
+from pymock_api.model import APIConfig
 from pymock_api.model.api_config import APIParameter as PyMockAPIParameter
 from pymock_api.model.api_config import BaseConfig, MockAPI, MockAPIs, _Config
 from pymock_api.model.enums import ResponseStrategy
