+++ test/unit_test/model/api_config/init.py
@@ -10,7 +10,6 @@ from pymock_api._utils import YAML
 from pymock_api._utils.file_opt import JSON
 from pymock_api.model.api_config import (
     HTTP,
-    APIConfig,
     APIParameter,
     BaseConfig,
     HTTPRequest,
@@ -33,8 +32,6 @@ from pymock_api.model.enums import Format, ResponseStrategy
 
 from ...._values import (
     _Base_URL,
-    _Config_Description,
-    _Config_Name,
     _Mock_Templatable_Setting,
     _Mock_Template_API_Request_Setting,
     _Mock_Template_API_Response_Setting,
@@ -44,7 +41,6 @@ from ...._values import (
     _Mock_Template_Setting,
     _Mock_Template_Values_Setting,
     _Test_API_Parameter,
-    _Test_Config,
     _Test_HTTP_Resp,
     _Test_Iterable_Parameter_Item_Name,
     _Test_Iterable_Parameter_Item_Value,
@@ -94,99 +90,6 @@ class TemplatableConfigTestSuite(ConfigTestSpec, ABC):
 
         # Verify
         assert serialized_sut.get("apply_template_props", None) is None
-
-
-class TestAPIConfig(ConfigTestSpec):
-    @pytest.fixture(scope="function")
-    def sut(self) -> APIConfig:
-        return APIConfig(name=_Config_Name, description=_Config_Description, apis=self._Mock_Model.mock_apis)
-
-    @pytest.fixture(scope="function")
-    def sut_with_nothing(self) -> APIConfig:
-        return APIConfig()
-
-    def test___len___with_value(self, sut: APIConfig):
-        sut.apis = _TestConfig.Mock_APIs
-        assert len(sut) == len(
-            list(filter(lambda k: k not in ["base", "template"], _TestConfig.Mock_APIs.keys()))
-        ), "The size of *APIConfig* data object should be same as *MockAPIs* data object."
-
-    def test___len___with_non_value(self):
-        assert len(APIConfig()) == 0, "The size of *APIConfig* data object should be '0' if property *apis* is None."
-
-    def test_has_apis_with_value(self, sut: APIConfig):
-        sut.apis = _TestConfig.Mock_APIs
-        assert sut.has_apis() is True, "It should be 'True' if its property *apis* has value."
-
-    def test_has_apis_with_no_value(self):
-        assert APIConfig().has_apis() is False, "It should be 'False' if its property *apis* is None."
-
-    def test_value_attributes(self, sut: APIConfig):
-        assert sut.name == _Config_Name, _assertion_msg
-        assert sut.description == _Config_Description, _assertion_msg
-        assert sut.apis.base.url == _TestConfig.API_Config.get("mocked_apis").get("base").get("url"), _assertion_msg
-        assert (
-            list(sut.apis.apis.keys())[0] in _TestConfig.API_Config.get("mocked_apis").get("apis").keys()
-        ), _assertion_msg
-
-    @pytest.mark.parametrize(
-        ("setting_val", "should_call_deserialize"),
-        [
-            ({}, False),
-            ({"test": "test"}, True),
-            (Mock(MockAPIs()), False),
-        ],
-    )
-    @patch.object(MockAPIs, "deserialize", return_value=MOCK_RETURN_VALUE)
-    def test_prop_apis_with_valid_obj(
-        self,
-        mock_deserialize: Mock,
-        setting_val: Union[dict, BaseConfig],
-        should_call_deserialize: bool,
-        sut_with_nothing: APIConfig,
-    ):
-        # Run target function
-        sut_with_nothing.apis = setting_val
-        # Verify the running result
-        if should_call_deserialize:
-            mock_deserialize.assert_called_once_with(data=setting_val)
-            assert sut_with_nothing.apis == MOCK_RETURN_VALUE
-        else:
-            mock_deserialize.assert_not_called()
-            assert sut_with_nothing.apis == (setting_val if setting_val else None)
-
-    @patch.object(MockAPIs, "deserialize", return_value=MOCK_RETURN_VALUE)
-    def test_prop_apis_with_invalid_obj(self, mock_deserialize: Mock, sut_with_nothing: APIConfig):
-        with pytest.raises(TypeError) as exc_info:
-            sut_with_nothing.apis = "Invalid object"
-        mock_deserialize.assert_not_called()
-        assert re.search(r"Setter .{1,32} only accepts .{1,32} type object.", str(exc_info.value), re.IGNORECASE)
-
-    def _expected_serialize_value(self) -> dict:
-        return _TestConfig.API_Config
-
-    def _expected_deserialize_value(self, obj: APIConfig) -> None:
-        assert isinstance(obj, APIConfig)
-        assert obj.name == _Config_Name
-        assert obj.description == _Config_Description
-        assert obj.apis.base.url == _TestConfig.API_Config.get("mocked_apis").get("base").get("url")
-        assert (
-            list(obj.apis.apis.keys())[0] in _TestConfig.API_Config.get("mocked_apis").get("apis").keys()
-        ), _assertion_msg
-
-    @patch("pymock_api._utils.file_opt.YAML.read", return_value=_TestConfig.API_Config)
-    def test_from_yaml_file(self, mock_read_yaml: Mock, sut: APIConfig):
-        config = sut.from_yaml(path="test-api.yaml")
-        mock_read_yaml.assert_called_once_with("test-api.yaml")
-        assert isinstance(config, APIConfig)
-        assert config.name == _Config_Name
-        assert config.description == _Config_Description
-        assert config.apis.serialize() == _TestConfig.Mock_APIs
-
-    @patch("pymock_api._utils.file_opt.YAML.write", return_value=None)
-    def test_to_yaml_file(self, mock_write_yaml: Mock, sut: APIConfig):
-        sut.to_yaml(path=_Test_Config)
-        mock_write_yaml.assert_called_once_with(path=_Test_Config, config=sut.serialize())
 
 
 class TestMockAPIs(ConfigTestSpec):
