+++ test/unit_test/model/swagger_config.py
@@ -8,7 +8,7 @@ from typing import List, Tuple
 
 import pytest
 
-from pymock_api.model.api_config import APIConfig
+from pymock_api.model import APIConfig
 from pymock_api.model.api_config import APIParameter as PyMockAPIParameter
 from pymock_api.model.api_config import MockAPI, _Config
 from pymock_api.model.enums import ResponseStrategy
